requirements.txt
fitz
numpy
requests
rank-bm25
scikit-learn
networkx
matplotlib
tqdm
Pillow
faiss-cpu
pandas
ipywidgets
rdflib
spacy
pyvis
datasets
transformers
accelerate
google-genai
PyMuPDF
python-dotenv
sentence-transformers